CM - Inclusão dos Scripts de criação de estrutura e carga de dados direto no projeto (para avaliação), ajustes após testes gerais (Models), alteração de nomenclatura para métodos assíncronos.

+++ BalcaoDeOfertasAPI/1 - Models/Moeda.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Nome { get; set; }
         public int QuantidadeTotal { get; set; }
-        public double ValorReal { get; set; }
+        public decimal ValorReal { get; set; }
 
         public Guid CarteiraId { get; set; }
         public Carteira Carteira { get; set; }
